DisplayGraphics provides access to painting mutable image using getImage method
git-svn-id: http://microemu.googlecode.com/svn/trunk@312 272dc960-a678-11de-8dbb-b5444cbab91d

+++ microemulator/src/com/barteo/emulator/app/capture/DeviceDisplayStream.java
@@ -105,7 +105,7 @@ public class DeviceDisplayStream implements PushBufferStream, DisplayRepaintList
 		synchronized (this) {
 			System.err.println("  - creating image: " + seqNo);
 
-			byte data[] = image.getData();
+			int data[] = image.getData();
 			buf.setData(data);
 
 			System.err.println("    read " + data.length + " bytes.");
